MatricesForHomalg

Documentation for MatricesForHomalg.

MatricesForHomalg.MatricesForHomalg

Base.:*

MatricesForHomalg.HomalgColumnVector

MatricesForHomalg.HomalgColumnVector

MatricesForHomalg.HomalgDiagonalMatrix

MatricesForHomalg.HomalgIdentityMatrix

MatricesForHomalg.HomalgMatrix

MatricesForHomalg.HomalgRowVector

MatricesForHomalg.HomalgRowVector

MatricesForHomalg.HomalgZeroMatrix

MatricesForHomalg.IsEmptyMatrix

MatricesForHomalg.IsOne

MatricesForHomalg.IsSymmetricMatrix

MatricesForHomalg.IsZero

MatricesForHomalg.NumberColumns

MatricesForHomalg.NumberRows

MatricesForHomalg.MatricesForHomalg — Module

Matrices for the homalg project

Base.:* — Method

R * mat

Rewrite the matrix mat over the ring R (if possible)

julia> mat = HomalgMatrix([1,2,3,4,5,6], 2, 3, ZZ)
[1   2   3]
[4   5   6]

julia> QQ * mat
[1//1   2//1   3//1]
[4//1   5//1   6//1]

julia> qmat = QQ * mat
[1//1   2//1   3//1]
[4//1   5//1   6//1]

julia> ZZ * qmat == mat
true

MatricesForHomalg.HomalgColumnVector — Method

HomalgColumnVector(L, r, R)

Construct a (r x 1)-matrix over the ring R with entries in the list L

julia> mat = HomalgColumnVector(1:5, 5, ZZ)
[1]
[2]
[3]
[4]
[5]

MatricesForHomalg.HomalgColumnVector — Method

HomalgColumnVector(L, r, R)

Construct a (r x 1)-matrix over the ring R with entries in the list L, where r = length(L)

julia> mat = HomalgColumnVector(1:5, ZZ)
[1]
[2]
[3]
[4]
[5]

MatricesForHomalg.HomalgDiagonalMatrix — Method

HomalgDiagonalMatrix(L, R)

Construct a diagonal matrix over the ring R using the list L of diagonal entries

julia> mat = HomalgDiagonalMatrix(1:5, ZZ)
[1   0   0   0   0]
[0   2   0   0   0]
[0   0   3   0   0]
[0   0   0   4   0]
[0   0   0   0   5]

MatricesForHomalg.HomalgIdentityMatrix — Method

HomalgIdentityMatrix(r, R)

Construct a (r x r)-identity matrix over the ring R

julia> mat = HomalgIdentityMatrix(3, ZZ)
[1   0   0]
[0   1   0]
[0   0   1]

MatricesForHomalg.HomalgMatrix — Method

HomalgMatrix(L, r, c, R)

Construct a (r x c)-matrix over the ring R with L as list of entries

julia> mat = HomalgMatrix([1,2,3,4,5,6], 2, 3, ZZ)
[1   2   3]
[4   5   6]

MatricesForHomalg.HomalgRowVector — Method

HomalgRowVector(L, c, R)

Construct a (1 x c)-matrix over the ring R with entries in the list L

julia> mat = HomalgRowVector(1:5, 5, ZZ)
[1   2   3   4   5]

MatricesForHomalg.HomalgRowVector — Method

HomalgRowVector(L, R)

Construct a (1 x c)-matrix over the ring R with entries in the list L, where c = length(L)

julia> mat = HomalgRowVector(1:5, ZZ)
[1   2   3   4   5]

MatricesForHomalg.HomalgZeroMatrix — Method

HomalgZeroMatrix(r, c, R)

Construct a (r x c)-zero matrix over the ring R

julia> mat = HomalgZeroMatrix(3, 2, ZZ)
[0   0]
[0   0]
[0   0]

MatricesForHomalg.IsEmptyMatrix — Method

IsEmptyMatrix(mat)

Return true if mat does not contain any entry, otherwise false

julia> mat = HomalgMatrix([], 0, 0, ZZ)
0 by 0 empty matrix

julia> IsEmptyMatrix(mat)
true

MatricesForHomalg.IsOne — Method

IsOne(mat)

Return true if mat is an identity matrix, otherwise false

julia> mat = HomalgIdentityMatrix(3, ZZ)
[1   0   0]
[0   1   0]
[0   0   1]

julia> IsOne(mat)
true

MatricesForHomalg.IsSymmetricMatrix — Method

IsSymmetricMatrix(mat)

Return true if the matrix mat is symmetric with respect to its main diagonal, otherwise false

julia> mat = HomalgMatrix([1,2,3,2,4,5,3,5,6], 3, 3, ZZ)
[1   2   3]
[2   4   5]
[3   5   6]

julia> IsSymmetricMatrix(mat)
true

MatricesForHomalg.IsZero — Method

IsZero(mat)

Return true if mat is a zero matrix, otherwise false

julia> mat = HomalgZeroMatrix(3, 2, ZZ)
[0   0]
[0   0]
[0   0]

julia> IsZero(mat)
true

MatricesForHomalg.NumberColumns — Method

NumberColumns(mat)

The number of columns of the matrix mat

julia> mat = HomalgMatrix(1:6, 2, 3, ZZ)
[1   2   3]
[4   5   6]

julia> NumberColumns(mat)
3

MatricesForHomalg.NumberRows — Method

NumberRows(mat)

The number of rows of the matrix mat

julia> mat = HomalgMatrix(1:6, 2, 3, ZZ)
[1   2   3]
[4   5   6]

julia> NumberRows(mat)
2