MatricesForHomalg

Documentation for MatricesForHomalg.

MatricesForHomalg.MatricesForHomalg

Base.:*

MatricesForHomalg.HomalgColumnVector

MatricesForHomalg.HomalgColumnVector

MatricesForHomalg.HomalgDiagonalMatrix

MatricesForHomalg.HomalgIdentityMatrix

MatricesForHomalg.HomalgMatrix

MatricesForHomalg.HomalgRowVector

MatricesForHomalg.HomalgRowVector

MatricesForHomalg.HomalgZeroMatrix

MatricesForHomalg.IsOne

MatricesForHomalg.IsZero

MatricesForHomalg.NrColumns

MatricesForHomalg.NrRows

MatricesForHomalg.MatricesForHomalg — Module

Matrices for the homalg project

Base.:* — Method

R * mat

Rewrite the matrix mat over the ring R (if possible)

MatricesForHomalg.HomalgColumnVector — Method

HomalgColumnVector(L, r, R)

Construct a (r x 1)-matrix over the ring R with entries in the list L

MatricesForHomalg.HomalgColumnVector — Method

HomalgColumnVector(L, r, R)

Construct a (r x 1)-matrix over the ring R with entries in the list L, where r = length(L)

MatricesForHomalg.HomalgDiagonalMatrix — Method

HomalgDiagonalMatrix(L, R)

Construct a diagonal matrix over the ring R using the list L of diagonal entries

MatricesForHomalg.HomalgIdentityMatrix — Method

HomalgIdentityMatrix(r, R)

Construct a (r x r)-identity matrix over the ring R

MatricesForHomalg.HomalgMatrix — Method

HomalgMatrix(L, r, c, R)

Construct a (r x c)-matrix over the ring R with L as list of entries

MatricesForHomalg.HomalgRowVector — Method

HomalgRowVector(L, c, R)

Construct a (1 x c)-matrix over the ring R with entries in the list L

MatricesForHomalg.HomalgRowVector — Method

HomalgRowVector(L, R)

Construct a (1 x c)-matrix over the ring R with entries in the list L, where c = length(L)

MatricesForHomalg.HomalgZeroMatrix — Method

HomalgZeroMatrix(r, c, R)

Construct a (r x c)-zero matrix over the ring R

MatricesForHomalg.IsOne — Method

IsOne(mat)

Return true if mat is an identity matrix, otherwise false

MatricesForHomalg.IsZero — Method

IsZero(mat)

Return true if mat is a zero matrix, otherwise false

MatricesForHomalg.NrColumns — Method

NrColumns(mat)

The number of columns of the matrix mat

MatricesForHomalg.NrRows — Method

NrRows(mat)

The number of rows of the matrix mat